requirements.txt
requests
flask
flask_socketio
python-dotenv